Rename UserHistoryDrawerAccountHistory to UserDrawerAccountHistory
Cleans up the agglutination's on these file names.

Probably means we should rename UserHistoryDrawer... to
UserDrawer... everywhere. Will get to that later.

CORL-234
